chore: move release PR lookup into its own file
The multi-pull-request support pushed release_pr_checks.go past the 500-line
production file limit. Finding and matching the pending release pull requests is
a self-contained concern, so it moves out whole rather than being trimmed.

Also drops a trailing period from a merge failure message that staticcheck
rejects.
